+++ xroad-catalog-persistence/src/main/java/fi/vrk/xroad/catalog/persistence/repository/ServiceRepository.java
@@ -115,6 +115,17 @@ public interface ServiceRepository extends CrudRepository<Service, Long> {
             +"AND s.subsystem.subsystemCode = :subsystemCode "
             + "AND s.subsystem.member.memberCode = :memberCode "
             + "AND s.subsystem.member.memberClass = :memberClass "
+            + "AND s.subsystem.member.xRoadInstance = :xRoadInstance ")
+    List<Service> findServicesByMemberServiceAndSubsystem(@Param("xRoadInstance") String xRoadInstance,
+                                                          @Param("memberClass") String memberClass,
+                                                          @Param("memberCode") String memberCode,
+                                                          @Param("serviceCode") String serviceCode,
+                                                          @Param("subsystemCode") String subsystemCode);
+
+    @Query("SELECT s FROM Service s WHERE s.serviceCode = :serviceCode "
+            +"AND s.subsystem.subsystemCode = :subsystemCode "
+            + "AND s.subsystem.member.memberCode = :memberCode "
+            + "AND s.subsystem.member.memberClass = :memberClass "
             + "AND s.subsystem.member.xRoadInstance = :xRoadInstance "
             +"AND s.serviceVersion = :serviceVersion")
     Service findAllByMemberServiceAndSubsystemAndVersion(@Param("xRoadInstance") String xRoadInstance,